ffi, sockets: add sock-opt and sock-set-opt.
The new sock-opt and sock-set-opt functions are wrappers around
getsockopt and setsockopt, respectively.

All POSIX socket options are registered. Platform-specific
options may be added in the future.

* ffi.c (sock_opt, sock_set_opt): New functions.
(ffi_init): Register sock-opt, sock-set-opt, sol-socket,
ipproto-ip, ipproto-ipv6, ipproto-tcp, ipproto-udp,
so-acceptconn, so-broadcast, so-debug, so-dontroute, so-error,
so-keepalive, so-linger, so-oobinline, so-rcvbuf, so-rcvlowat,
so-rcvtimeo, so-reuseaddr, so-sndbuf, so-sndlowat, so-sndtimeo,
so-type, ipv6-join-group, ipv6-leave-group, ipv6-multicast-hops,
ipv6-multicast-if, ipv6-multicast-loop, ipv6-unicast-hops,
ipv6-v6only, tcp-nodelay.

* lisplib.c (sock_set_entries): Add sock-opt and sock-set-opt.

* stdlib/socket.tl (sock-opt): Define as syntactic place.

* tests/014/socket-misc.tl: New cases, for sock-opt.
(set-and-get): New macro.

* txr.1: Documented. Also, mention that sock-bind enables
so-reuseaddr.

* stdlib/doc-syms.tl: Updated.