New variable: *child-env*.
This specifies the environment to be used for executing
programs.

* stream.c (open_subprocess, run): Check *child-env* variable and if
other than t, then install the environment before execvp.
In the spawn-based version of run, we save and restore the
environment around the spawn call, if *child-env* is in
effect.

* sysif.c (child_env_s): New symbol variable.
(exec_wrap): If *child-env* is other than t, then save the
environment in a list, and install the specified environment
before calling execvp. If that function returns, restore the
environbment.

* sysif.h (child_env_s): Declared.
(child_env): New macro.

* tests/018/process.tl: New tests.

* txr.1: Documented.

* stdlib/doc-syms.tl: Updated.