parser: bugfix: #; at front of list.
The parser wrongly reads #(#; abc) as (nil) instead of (), and
related cases derived from this one are all likewise wrong.
A number of tests added in the previous commit target this
and fail. They are hereby fixed.

* parser.y (listacc): In the productions that begin with
HASH_SEMI, do not produce a (nil . nil) leading cons, but a
(nao . nil) leading cons; so the fact that the first item is
commented out is represented by a nao in the car field of the
leading cons.
(n_exprs): If the first element of the list produced by the
listacc grammar symbol is nao, then pop it off.  Thereby, we
lose the spurious nil that we previously had there left by the
commented-out item.

* y.tab.c.shipped: Updated.