compiler: also clear .next before re-linking graph.
* stdlib/optimize.tl (basic-blocks elim-dead-code): When
clearing the links before recalculating the graph, also clear
the next field of every block, because link-graph only sets
this if necessary, assuming that the value is already nil.
Thus by not resetting it, we risk leaving stale values in
these .next fields.  The code reachability calculation relies
on next fields, so if they falsely point to dead blocks, those
blocks could be falsely retained.