less: symbolic arguments: fix crash and incorrectness.
* lib.c (less): We cannot direclty access right->s.package
because the right operand can be nil. This causes a crash.
Furthermore, the separate NIL case is wrong. If the left
object is nil, the same logic must be carried out as for SYM.
The opposite operand might have the same name, and so packages
have to be compared. We simply merge the two cases, and make
sure we use the proper accessors symbol_name and
symbol_package to avoid blowing up on nil.