compiler: catch bugfix.
Commit c8f12ee44d226924b89cdd764b65a5f6a4030b81 tried to fix
an aspect of this problem. I ran into an issue where the try
code produced a D register as its output, and this was
clobbered by the catch code.  In fact, the catch code simply
must not clobber the try fragment's output register. No matter
what register that is, it is not safe. A writable T register
could hold a variable.

For instance, this infinitely looping code is miscompiled
such that it terminates:

  (let ((x 42))
    (while (eql x 42)
      (catch
        (progn (throw 'foo)
               x)
        (foo () 0))))

When the exception is caught by the (foo () 0) clause
x is overwritten with that 0 value.

The variable x is assigned to a register like t13,
and since the progn form returns x as it value, it
compiles to a fragment (tfrag) which indicates t13
as its output register.

The catch code wrongly borrows ohis as its own output
register, placing the 0 value into it.

* stdlib/compiler.tl (compiler comp-catch): Get rid of the
coreg local variable, replacing all its uses with oreg.