compiler: new late-peephole case.
This is related to the pattern in the previous commit. When we have a
situation like this:

 lab1
   mov tn nil
 lab2
   ifq tn nil lab4
 lab3
   gcall tn ...

We know that if lab1 is entered, then lab2 will necessarily
fall through: the lab4 branch is not taken because tn is nil.
But then, tn is clobbered immediately in lab3 by the gcall tn.
In other words, the value stored into tn by lab1 is never used.

Therefore, we can remove the "mov tn nil" instruction and
move the l1 label.

 lab2
   ifq tn nil lab4
 lab1
 lab3
   gcall tn ...

There are 74 hits for this pattern in stdlib.

* stdlib/optimize.tl (basic-blocks late-peephole): Implement the
above pattern.