maprodo: bugfix: spurious return value.
There are cases when maprodo returns a non-nil value, even
though it is supposed to collect nothing. This is because
though it is is collecting nothing, that nothing is sometimes
converted to an alternative return type via make_like.

* eval.c (prod_common): We allow the collect_fn function
pointer to be null, to indicate nothing is to be collected,
rather than using a stub. If collect_fn is null, we just call
the mapping function without collecting its value, and at the
end, we do not involve make_like and just return nil.
(collect_nothing): Static function removed.
(maprodo): Pass null function pointer instead of collect_nothing.