configure: broken file offset bits detection.
We are not detecting the need to do -D_FILE_OFFSET_BITS=64
correct, resulting in no large file support on 32 bit
platforms based on Glibc. This is a regression since TXR 244.

* configure: We must pass EXTRA_FLAGS=-D$try to
actually try the options we are looping over.  This argument
was accidentally removed in commit
3d80caccafc27ac812bbf8226eba6d8e529c63ff
on October 9, 2020, when the conftest_symns command was
changed to conftest_o.