sysif: rlim constants: use signed/unsigned-agnostic macro.
* lib.h (num_ex): New macro. Uses unum if the argument is out
of range for the signed type. Thus we can use this with
unsigned constants that would wrap negative if passed to num.
This is useful if some type in a system header file might be
signed or unsigned.

* sysif.c (sysif_init): Use num_ex for the RLIM_* constants.
I'm observing values of -1 which should really be large,
positive values in the rlim_t type, that being unsigned.