freeform: bug: account for consumed prefix.
This was reported by user vapnik spaknik. The @(freeform),
when reconstituting the unmatched trailing portion of the
virtual line back into a list of lines, uses the abstract
match position, neglecting to account for the fact that a
prefix of the line may have been physically consumed to save
memory.

* match.c (v_freeform): When calling
lazy_str_get_trailing_list, indicate the correct amount of
prefix material, by subtracting, from the matching length,
the base variable, which indicates how much of the prefix had
been consumed. This consumption takes place above 4000 bytes,
which is why the freeform test cases are not catching this.

* tests/006/freeform-5.txr: New file.

* tests/006/freeform-5.expected: New file.